productoAgregar

+++ SacNew/Models/LocacionKilometrosEntre.cs
@@ -5,7 +5,7 @@
         public int IdKilometros { get; set; }
         public int IdLocacionOrigen { get; set; }
         public int IdLocacionDestino { get; set; }
-        public int Kilometros { get; set; }
+        public decimal Kilometros { get; set; }
 
         // Relación con locaciones
         public Locacion LocacionDestino { get; set; }  // Locacion destino
